Move to "docs/reference-implementation"
Don't put it into the repository root, it's cluttered as it is now.

The new name also makes it explicit that it's not "reference" in the
documentation sense, but some code. Hopefully. We're stuck with this
name from now on.

Similarly, put integration tests into "tests/reference-implementation".
