+++ docs/reference-implementation/go.mod
@@ -1,0 +1,3 @@
+module github.com/cossacklabs/themis/docs/reference-implementation
+
+go 1.16
