+++ docs/reference-implementation/soter/soter-alg.go
