+++ docs/reference-implementation/soter/soter-container.go
