--- reference/go.mod
@@ -1,3 +1,0 @@
-module github.com/cossacklabs/themis/reference
-
-go 1.16
