+++ tests/reference-implementation/go.mod
@@ -1,0 +1,3 @@
+module reference-implementation-integration-tests
+
+go 1.16
