+++ tests/reference-implementation/soter/container_test.go
@@ -3,8 +3,8 @@ package soter
 import (
 	"testing"
 
+	"github.com/cossacklabs/themis/docs/reference-implementation/soter"
 	"github.com/cossacklabs/themis/gothemis/keys"
-	"github.com/cossacklabs/themis/reference/soter"
 )
 
 // Asymmetric keys generated by GoThemis are packed into Soter containers.
