--- tests/reference/go.mod
@@ -1,3 +1,0 @@
-module reference-integration-tests
-
-go 1.16
